+++ src/composer.rs
@@ -105,10 +105,10 @@ pub trait Composer: Sized + Index<Witness, Output = BlsScalar> {
         self.append_custom_gate_internal(constraint)
     }
 
-    /// Performs a logical AND or XOR op between the inputs provided for the
-    /// specified number of bits (counting from the least significant bit).
+    /// Performs a logical AND or XOR op between the inputs provided for
+    /// `num_bits = BIT_PAIRS * 2` bits (counting from the least significant).
     ///
-    /// Each logic gate adds `(num_bits / 2) + 1` gates to the circuit to
+    /// Each logic gate adds `BIT_PAIRS + 1` gates to the circuit to
     /// perform the whole operation.
     ///
     /// ## Constraint
@@ -116,22 +116,15 @@ pub trait Composer: Sized + Index<Witness, Output = BlsScalar> {
     ///   `a` and `b`.
     /// - is_component_xor = 0 -> Performs AND between the first `num_bits` for
     ///   `a` and `b`.
-    ///
-    /// # Panics
-    /// This function will panic if the num_bits specified is not even, ie.
-    /// `num_bits % 2 != 0`.
-    fn append_logic_component(
+    fn append_logic_component<const BIT_PAIRS: usize>(
         &mut self,
         a: Witness,
         b: Witness,
-        num_bits: usize,
         is_component_xor: bool,
     ) -> Witness {
         // the bits are iterated as chunks of two; hence, we require an even
         // number
-        debug_assert_eq!(num_bits & 1, 0);
-
-        let num_bits = cmp::min(num_bits, 256);
+        let num_bits = cmp::min(BIT_PAIRS * 2, 256);
         let num_quads = num_bits >> 1;
 
         let bls_four = BlsScalar::from(4u64);
@@ -272,7 +265,13 @@ pub trait Composer: Sized + Index<Witness, Output = BlsScalar> {
         let width = 2;
         let wnaf_entries = scalar.compute_windowed_naf(width);
 
-        debug_assert_eq!(wnaf_entries.len(), bits);
+        // this will pass as long as `compute_windowed_naf` returns an array
+        // with 256 elements
+        debug_assert_eq!(
+            wnaf_entries.len(),
+            bits,
+            "the wnaf_entries array is expected to be 256 elements long"
+        );
 
         // initialize the accumulators
         let mut scalar_acc = vec![BlsScalar::zero()];
@@ -581,35 +580,25 @@ pub trait Composer: Sized + Index<Witness, Output = BlsScalar> {
     }
 
     /// Adds a logical AND gate that performs the bitwise AND between two values
-    /// for the specified first `num_bits` returning a [`Witness`]
+    /// specified first `num_bits = BIT_PAIRS * 2` bits returning a [`Witness`]
     /// holding the result.
-    ///
-    /// # Panics
-    ///
-    /// If the `num_bits` specified in the fn params is odd.
-    fn append_logic_and(
+    fn append_logic_and<const BIT_PAIRS: usize>(
         &mut self,
         a: Witness,
         b: Witness,
-        num_bits: usize,
     ) -> Witness {
-        self.append_logic_component(a, b, num_bits, false)
+        self.append_logic_component::<BIT_PAIRS>(a, b, false)
     }
 
     /// Adds a logical XOR gate that performs the XOR between two values for the
-    /// specified first `num_bits` returning a [`Witness`] holding the
-    /// result.
-    ///
-    /// # Panics
-    ///
-    /// If the `num_bits` specified in the fn params is odd.
-    fn append_logic_xor(
+    /// specified first `num_bits = BIT_PAIRS * 2` bits returning a [`Witness`]
+    /// holding the result.
+    fn append_logic_xor<const BIT_PAIRS: usize>(
         &mut self,
         a: Witness,
         b: Witness,
-        num_bits: usize,
     ) -> Witness {
-        self.append_logic_component(a, b, num_bits, true)
+        self.append_logic_component::<BIT_PAIRS>(a, b, true)
     }
 
     /// Constrain `a` to be equal to `constant + pi`.
@@ -913,18 +902,18 @@ pub trait Composer: Sized + Index<Witness, Output = BlsScalar> {
     }
 
     /// Adds a range-constraint gate that checks and constrains a [`Witness`]
-    /// to be encoded in at most `num_bits`, which means that it will be within
-    /// the range `[0, 2^num_bits[`.
+    /// to be encoded in at most `num_bits = BIT_PAIRS * 2` bits, which means
+    /// that the underlying [`BlsScalar`] of the [`Witness`] will be within the
+    /// range `[0, 2^num_bits[`, where `num_bits` is dividable by two.
     ///
-    /// This function adds min(1, `num_bits/4`) gates to the circuit description
-    /// in order to add the range constraint.
-    ///
-    ///# Panics
-    /// This function will panic if the num_bits specified is not even, ie.
-    /// `num_bits % 2 != 0`.
-    fn component_range(&mut self, witness: Witness, num_bits: usize) {
-        // number of bits must be even
-        debug_assert_eq!(num_bits % 2, 0);
+    /// This function adds:
+    /// (num_bits - 1)/8 + 9 gates, when num_bits > 0,
+    /// and 7 gates, when num_bits = 0
+    /// to the circuit description.
+    fn component_range<const BIT_PAIRS: usize>(&mut self, witness: Witness) {
+        // the bits are iterated as chunks of two; hence, we require an even
+        // number
+        let num_bits = cmp::min(BIT_PAIRS * 2, 256);
 
         // if num_bits = 0 constrain witness to 0
         if num_bits == 0 {