+++ tests/composer.rs
@@ -64,7 +64,7 @@ fn circuit_with_all_gates() {
             composer.gate_add(Constraint::new().left(1).right(1).a(w_a).b(w_b));
 
             composer.component_add_point(w_z, w_z);
-            composer.append_logic_and(w_a, w_b, 254);
+            composer.append_logic_and::<128>(w_a, w_b);
             composer.component_boolean(Builder::ONE);
             composer.component_decomposition::<254>(w_a);
             composer.component_mul_generator(
@@ -72,13 +72,13 @@ fn circuit_with_all_gates() {
                 dusk_jubjub::GENERATOR_EXTENDED,
             )?;
             composer.component_mul_point(w_y, w_z);
-            composer.component_range(w_a, 254);
+            composer.component_range::<128>(w_a);
             composer.component_select(Builder::ONE, w_a, w_b);
             composer.component_select_identity(Builder::ONE, w_z);
             composer.component_select_one(Builder::ONE, w_a);
             composer.component_select_point(Builder::ONE, w_z, w_z);
             composer.component_select_zero(Builder::ONE, w_a);
-            composer.append_logic_xor(w_a, w_b, 254);
+            composer.append_logic_xor::<128>(w_a, w_b);
 
             Ok(())
         }