+++ tests/logic.rs
@@ -4,39 +4,40 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use core::cmp;
 use dusk_plonk::prelude::*;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
 
 mod common;
-use common::{check_satisfied_circuit, check_unsatisfied_circuit, setup};
+use common::{check_satisfied_circuit, check_unsatisfied_circuit};
 
 #[test]
 fn append_logic_and() {
     #[derive(Default)]
-    pub struct TestCircuit {
+    pub struct TestCircuit<const BIT_PAIRS: usize> {
         a: BlsScalar,
         b: BlsScalar,
         result: BlsScalar,
-        bits: usize,
     }
 
-    impl TestCircuit {
-        pub fn new(a: BlsScalar, b: BlsScalar, bits: usize) -> Self {
+    impl<const BIT_PAIRS: usize> TestCircuit<BIT_PAIRS> {
+        pub fn new(a: BlsScalar, b: BlsScalar) -> Self {
+            let bits = cmp::min(BIT_PAIRS * 2, 256);
             let bit_mask = BlsScalar::pow_of_2(bits as u64) - BlsScalar::one();
 
-            // BlsScalar are max 255 bits long so a bit_mask with more than 255
-            // bits will be overflowing and therefore incorrect
-            let result = match bits < 256 {
+            // BlsScalar are max 255 bits long which means that a bit_mask with
+            // more than 255 bits will be overflowing and therefore incorrect
+            let result = match bits <= 255 {
                 true => a & b & bit_mask,
                 false => a & b,
             };
 
-            Self { a, b, result, bits }
+            Self { a, b, result }
         }
     }
 
-    impl Circuit for TestCircuit {
+    impl<const BIT_PAIRS: usize> Circuit for TestCircuit<BIT_PAIRS> {
         fn circuit<C>(&self, composer: &mut C) -> Result<(), Error>
         where
             C: Composer,
@@ -45,7 +46,8 @@ fn append_logic_and() {
             let w_b = composer.append_witness(self.b);
             let w_result = composer.append_witness(self.result);
 
-            let circuit_result = composer.append_logic_and(w_a, w_b, self.bits);
+            let circuit_result =
+                composer.append_logic_and::<BIT_PAIRS>(w_a, w_b);
 
             composer.assert_equal(w_result, circuit_result);
 
@@ -58,8 +60,10 @@ fn append_logic_and() {
     let label = b"append_logic_and";
     let rng = &mut StdRng::seed_from_u64(0x1ead);
     let capacity = 1 << 8;
-    let (prover, verifier) =
-        setup(capacity, rng, label, &TestCircuit::default());
+    let pp = PublicParameters::setup(capacity, rng)
+        .expect("Creation of public parameter shouldn't fail");
+    let (prover, verifier) = Compiler::compile::<TestCircuit<0>>(&pp, label)
+        .expect("Circuit should compile");
 
     // Common public input vector to be used by all tests
     let pi = vec![];
@@ -68,65 +72,62 @@ fn append_logic_and() {
     //
     // Test default works
     let msg = "Default circuit verification should pass";
-    let circuit = TestCircuit::default();
+    let circuit = TestCircuit::<0>::default();
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test comparing 0 bits is always zero
     let msg = "Circuit verification of satisfied circuit should pass";
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
-    let circuit = TestCircuit {
+    let circuit = TestCircuit::<0> {
         a,
         b,
         result: BlsScalar::zero(),
-        bits: 0,
     };
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test with bits = 32
     //
     // Create new prover and verifier circuit descriptions
-    let bits = 32;
-    let a = BlsScalar::zero();
-    let b = BlsScalar::zero();
-    let circuit = TestCircuit::new(a, b, bits);
-    let (prover, verifier) = setup(capacity, rng, label, &circuit);
+    const BIT_PAIRS_16: usize = 16;
+    let (prover, verifier) =
+        Compiler::compile::<TestCircuit<BIT_PAIRS_16>>(&pp, label)
+            .expect("Circuit should compile");
 
     // Test sanity:
     let a = BlsScalar::from(0x0f0f_0ff0_0f0f_0ff0);
     let b = BlsScalar::from(0xffff_0000_0000_ffff);
     let result = BlsScalar::from(0x0000_0ff0);
-    let circuit = TestCircuit { a, b, result, bits };
+    let circuit: TestCircuit<BIT_PAIRS_16> = TestCircuit { a, b, result };
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test random works:
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
-    let circuit = TestCircuit::new(a, b, bits);
+    let circuit: TestCircuit<BIT_PAIRS_16> = TestCircuit::new(a, b);
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test invalid circuit fails
     let msg = "Proof creation of unsatisfied circuit should fail";
-    let bit_mask = BlsScalar::pow_of_2(bits as u64) - BlsScalar::one();
+    let bit_mask =
+        BlsScalar::pow_of_2(BIT_PAIRS_16 as u64 * 2) - BlsScalar::one();
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
     let right_result = a & b & bit_mask;
     let c = BlsScalar::random(rng);
     let wrong_result = a & c & bit_mask;
     assert_ne!(right_result, wrong_result);
-    let circuit_unsatisfied = TestCircuit {
+    let circuit_unsatisfied: TestCircuit<BIT_PAIRS_16> = TestCircuit {
         a,
         b,
         result: wrong_result,
-        bits,
     };
     check_unsatisfied_circuit(&prover, &circuit_unsatisfied, rng, &msg);
     // sanity check
-    let circuit_satisfied = TestCircuit {
+    let circuit_satisfied: TestCircuit<BIT_PAIRS_16> = TestCircuit {
         a,
         b,
         result: right_result,
-        bits,
     };
     check_satisfied_circuit(
         &prover,
@@ -140,24 +141,23 @@ fn append_logic_and() {
     // Test with bits = 256
     //
     // Create new circuit description for the prover and verifier
-    let bits = 256;
-    let a = BlsScalar::zero();
-    let b = BlsScalar::zero();
-    let circuit = TestCircuit::new(a, b, bits);
-    let (prover, verifier) = setup(capacity, rng, label, &circuit);
+    const BIT_PAIRS_128: usize = 128;
+    let (prover, verifier) =
+        Compiler::compile::<TestCircuit<BIT_PAIRS_128>>(&pp, label)
+            .expect("Circuit should compile");
 
     // Test sanity:
     let a = -BlsScalar::one();
     let b = -BlsScalar::one();
     let result = -BlsScalar::one();
-    let circuit = TestCircuit { a, b, result, bits };
+    let circuit: TestCircuit<BIT_PAIRS_128> = TestCircuit { a, b, result };
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test random works:
     let msg = "Circuit verification with random values should pass";
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
-    let circuit = TestCircuit::new(a, b, bits);
+    let circuit: TestCircuit<BIT_PAIRS_128> = TestCircuit::new(a, b);
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test invalid circuit fails
@@ -168,19 +168,17 @@ fn append_logic_and() {
     let c = BlsScalar::random(rng);
     let wrong_result = a & c;
     assert_ne!(right_result, wrong_result);
-    let circuit = TestCircuit {
+    let circuit: TestCircuit<BIT_PAIRS_128> = TestCircuit {
         a,
         b,
         result: wrong_result,
-        bits,
     };
     check_unsatisfied_circuit(&prover, &circuit, rng, &msg);
     // sanity check
-    let circuit_satisfied = TestCircuit {
+    let circuit_satisfied: TestCircuit<BIT_PAIRS_128> = TestCircuit {
         a,
         b,
         result: right_result,
-        bits,
     };
     check_satisfied_circuit(
         &prover,
@@ -190,36 +188,20 @@ fn append_logic_and() {
         rng,
         &"Sanity check should pass",
     );
-
-    // Test with odd bits = 55
-    //
-    // Create new prover and verifier circuit descriptions
-    let bits = 55;
-    let a = BlsScalar::zero();
-    let b = BlsScalar::zero();
-    let circuit = TestCircuit::new(a, b, bits);
-    let (prover, _) = setup(capacity, rng, label, &circuit);
-
-    // Test random fails:
-    let msg = "Circuit verification with odd bits should fail";
-    let a = BlsScalar::random(rng);
-    let b = BlsScalar::random(rng);
-    let circuit = TestCircuit::new(a, b, bits);
-    check_unsatisfied_circuit(&prover, &circuit, rng, &msg);
 }
 
 #[test]
 fn append_logic_xor() {
     #[derive(Default)]
-    pub struct TestCircuit {
+    pub struct TestCircuit<const BIT_PAIRS: usize> {
         a: BlsScalar,
         b: BlsScalar,
         result: BlsScalar,
-        bits: usize,
     }
 
-    impl TestCircuit {
-        pub fn new(a: BlsScalar, b: BlsScalar, bits: usize) -> Self {
+    impl<const BIT_PAIRS: usize> TestCircuit<BIT_PAIRS> {
+        pub fn new(a: BlsScalar, b: BlsScalar) -> Self {
+            let bits = cmp::min(BIT_PAIRS * 2, 256);
             let bit_mask = BlsScalar::pow_of_2(bits as u64) - BlsScalar::one();
 
             // BlsScalar are max 255 bits long so a bit_mask with more than 255
@@ -229,11 +211,11 @@ fn append_logic_xor() {
                 false => a ^ b,
             };
 
-            Self { a, b, result, bits }
+            Self { a, b, result }
         }
     }
 
-    impl Circuit for TestCircuit {
+    impl<const BIT_PAIRS: usize> Circuit for TestCircuit<BIT_PAIRS> {
         fn circuit<C>(&self, composer: &mut C) -> Result<(), Error>
         where
             C: Composer,
@@ -242,7 +224,8 @@ fn append_logic_xor() {
             let w_b = composer.append_witness(self.b);
             let w_result = composer.append_witness(self.result);
 
-            let circuit_result = composer.append_logic_xor(w_a, w_b, self.bits);
+            let circuit_result =
+                composer.append_logic_xor::<BIT_PAIRS>(w_a, w_b);
 
             composer.assert_equal(w_result, circuit_result);
 
@@ -255,8 +238,10 @@ fn append_logic_xor() {
     let label = b"append_logic_xor";
     let rng = &mut StdRng::seed_from_u64(0xdea1);
     let capacity = 1 << 8;
-    let (prover, verifier) =
-        setup(capacity, rng, label, &TestCircuit::default());
+    let pp = PublicParameters::setup(capacity, rng)
+        .expect("Creation of public parameter shouldn't fail");
+    let (prover, verifier) = Compiler::compile::<TestCircuit<0>>(&pp, label)
+        .expect("Circuit should compile");
 
     // Common values to be used by all tests
     let pi = vec![];
@@ -265,65 +250,63 @@ fn append_logic_xor() {
     //
     // Test default works
     let msg = "Default circuit verification should pass";
-    let circuit = TestCircuit::default();
+    let circuit = TestCircuit::<0>::default();
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test comparing 0 bits is always zero
     let msg = "Circuit verification of satisfied circuit should pass";
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
-    let circuit = TestCircuit {
+    let circuit: TestCircuit<0> = TestCircuit {
         a,
         b,
         result: BlsScalar::zero(),
-        bits: 0,
     };
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test with bits = 32
     //
     // Create new prover and verifier circuit descriptions
-    let bits = 32;
-    let a = BlsScalar::zero();
-    let b = BlsScalar::zero();
-    let circuit = TestCircuit::new(a, b, bits);
-    let (prover, verifier) = setup(capacity, rng, label, &circuit);
+    const BIT_PAIRS_16: usize = 16;
+    let (prover, verifier) =
+        Compiler::compile::<TestCircuit<BIT_PAIRS_16>>(&pp, label)
+            .expect("Circuit should compile");
 
     // Test sanity:
     let a = BlsScalar::from(0x0f0f_0ff0_0f0f_0ff0);
     let b = BlsScalar::from(0xffff_0000_0000_ffff);
     let result = BlsScalar::from(0x0f0f_f00f);
-    let circuit = TestCircuit { a, b, result, bits };
+    let circuit: TestCircuit<BIT_PAIRS_16> = TestCircuit { a, b, result };
+
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test random works:
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
-    let circuit = TestCircuit::new(a, b, bits);
+    let circuit: TestCircuit<BIT_PAIRS_16> = TestCircuit::new(a, b);
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test invalid circuit fails
     let msg = "Proof creation of unsatisfied circuit should fail";
-    let bit_mask = BlsScalar::pow_of_2(bits as u64) - BlsScalar::one();
+    let bit_mask =
+        BlsScalar::pow_of_2(BIT_PAIRS_16 as u64 * 2) - BlsScalar::one();
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
     let right_result = (a ^ b) & bit_mask;
     let c = BlsScalar::random(rng);
     let wrong_result = (a ^ c) & bit_mask;
     assert_ne!(right_result, wrong_result);
-    let circuit_unsatisfied = TestCircuit {
+    let circuit_unsatisfied: TestCircuit<BIT_PAIRS_16> = TestCircuit {
         a,
         b,
         result: wrong_result,
-        bits,
     };
     check_unsatisfied_circuit(&prover, &circuit_unsatisfied, rng, &msg);
     // sanity check
-    let circuit_satisfied = TestCircuit {
+    let circuit_satisfied: TestCircuit<BIT_PAIRS_16> = TestCircuit {
         a,
         b,
         result: right_result,
-        bits,
     };
     check_satisfied_circuit(
         &prover,
@@ -337,24 +320,23 @@ fn append_logic_xor() {
     // Test with bits = 256
     //
     // Create new prover and verifier circuit descriptions
-    let bits = 256;
-    let a = BlsScalar::zero();
-    let b = BlsScalar::zero();
-    let circuit = TestCircuit::new(a, b, bits);
-    let (prover, verifier) = setup(capacity, rng, label, &circuit);
+    const BIT_PAIRS_128: usize = 128;
+    let (prover, verifier) =
+        Compiler::compile::<TestCircuit<BIT_PAIRS_128>>(&pp, label)
+            .expect("Circuit should compile");
 
     // Test sanity:
     let a = -BlsScalar::one();
     let b = BlsScalar::zero();
     let result = -BlsScalar::one();
-    let circuit = TestCircuit { a, b, result, bits };
+    let circuit: TestCircuit<BIT_PAIRS_128> = TestCircuit { a, b, result };
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test random works:
     let msg = "Circuit verification with random values should pass";
     let a = BlsScalar::random(rng);
     let b = BlsScalar::random(rng);
-    let circuit = TestCircuit::new(a, b, bits);
+    let circuit: TestCircuit<BIT_PAIRS_128> = TestCircuit::new(a, b);
     check_satisfied_circuit(&prover, &verifier, &pi, &circuit, rng, &msg);
 
     // Test invalid circuit fails
@@ -365,19 +347,17 @@ fn append_logic_xor() {
     let c = BlsScalar::random(rng);
     let wrong_result = a ^ c;
     assert_ne!(right_result, wrong_result);
-    let circuit = TestCircuit {
+    let circuit: TestCircuit<BIT_PAIRS_128> = TestCircuit {
         a,
         b,
         result: wrong_result,
-        bits,
     };
     check_unsatisfied_circuit(&prover, &circuit, rng, &msg);
     // sanity check
-    let circuit_satisfied = TestCircuit {
+    let circuit_satisfied: TestCircuit<BIT_PAIRS_128> = TestCircuit {
         a,
         b,
         result: right_result,
-        bits,
     };
     check_satisfied_circuit(
         &prover,
@@ -387,20 +367,4 @@ fn append_logic_xor() {
         rng,
         &"Sanity check should pass",
     );
-
-    // Test with odd bits = 55
-    //
-    // Create new prover and verifier circuit descriptions
-    let bits = 55;
-    let a = BlsScalar::zero();
-    let b = BlsScalar::zero();
-    let circuit = TestCircuit::new(a, b, bits);
-    let (prover, _) = setup(capacity, rng, label, &circuit);
-
-    // Test random fails:
-    let msg = "Circuit verification with odd bits should fail";
-    let a = BlsScalar::random(rng);
-    let b = BlsScalar::random(rng);
-    let circuit = TestCircuit::new(a, b, bits);
-    check_unsatisfied_circuit(&prover, &circuit, rng, &msg);
 }